Remove bazel in favor of plain old webpack
I tried to run this using the provided Dockerfile but got
an error that I need to create a buildbuddy.io account.

This commit drops Bazel from this project altogether and
switches over to Webpack.

Disclaimer: I didn't write any of this; I just prompted
claude to "switch this project from bazel to webpack".